# spec/openapi.yaml
openapi: 3.0.1
info:
  title: ZapMe
  description: Open source application to control things
  termsOfService: https://zapme.app/tos
  contact:
    name: Contact me
    url: https://zapme.app/contact
  license:
    name: 'The Apache License, Version 2.0'
    url: https://zapme.app/license
  version: v1
servers:
  - url: https://api.zapme.app
paths:
...
  /api/v1/auth/signin:
    post:
      tags:
        - Auth
      summary: ''
      operationId: AuthSignIn
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/AuthSignIn'
              description: Message sent to server to authenticate user using username and password
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '413':
          description: Client Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
components:
  schemas:
...
    AuthSignIn:
      required:
        - password
        - rememberMe
        - usernameOrEmail
      type: object
      properties:
        usernameOrEmail:
          type: string
          description: Username or email address
          example: MyUsername
        password:
          type: string
          description: Password
          example: Hq2yP1B^Fho&zRHxHkEu
        rememberMe:
          type: boolean
          description: Make this login persist for a longer period of time
      additionalProperties: false
      description: Message sent to server to authenticate user using username and password
...
    ErrorDetails:
      required:
        - code
        - detail
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        suggestion:
          type: string
          nullable: true
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
            nullable: true
          nullable: true
        notification:
          allOf:
            - $ref: '#/components/schemas/UserNotification'
          nullable: true
      additionalProperties: false
...
    NotificationSeverityLevel:
      enum:
        - info
        - warning
        - error
      type: string
...
    SessionDto:
      required:
        - expiresAtUtc
        - issuedAtUtc
        - sessionToken
      type: object
      properties:
        sessionToken:
          type: string
          format: uuid
        issuedAtUtc:
          type: string
          format: date-time
        expiresAtUtc:
          type: string
          format: date-time
      additionalProperties: false
...
    UserNotification:
      required:
        - content
        - severity
      type: object
      properties:
        severity:
          allOf:
            - $ref: '#/components/schemas/NotificationSeverityLevel'
        content:
          type: string
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme.\r\n\r\n\r\n                      Enter 'Bearer' [space] and then your token in the text input below.\r\n                      \r\n\r\n                        Example: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]